Split AIanalysis workflows and updated related scripts
AIanalysis was split into two workflows to make it easier to re-run analysis based on different config
parameters related to heterozygous RNA reads and number of heterozygotes to include in analysis. In
conjunction to this, resulting output files were renamed with the config parameters. Rules for getting
rsids and adding them to the resulting DESeq2 output were also added.